refactor: RadioGroup 공통 컴포넌트 Components > Common > RadioGroup으로 이동
