+++ front/src/Components/Common/RadioGroup/index.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from "react";
 import { styled } from "styled-components";
-import useRadio from "../../hooks/useRadio";
+import useRadio from "../../../hooks/useRadio";
 
 const RadioGroup = ({
   title,
